Added validation gates to limit update and delete to records owned be the signed-in user. FINISHED

+++ kita-src/app/Http/Controllers/Logic/UserController.php
@@ -9,6 +9,7 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\RedirectResponse;
 use App\Http\Requests\StoreUserRequest;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 
 
@@ -16,6 +17,8 @@ class UserController extends Controller
 {
     public function index() 
     {
+        Gate::authorize('admin-only');
+        
         $users = User::all();
         
         return response()->json([
@@ -41,6 +44,8 @@ class UserController extends Controller
 
     public function update(StoreUserRequest $request, User $user) 
     {
+        Gate::authorize('manage-owned', $user);
+
         $validated = $request->validated();
 
         if ($request->filled('password')){
@@ -56,6 +61,8 @@ class UserController extends Controller
 
     public function destroy(User $user) 
     {
+        Gate::authorize('manage-owned', $user);
+
         $user->delete();
 
         return response()->json([
